requirements.txt
pandas>=2.2
pyarrow>=15.0
pydantic>=2.7
pyyaml>=6.0
python-dotenv>=1.0
pytest>=8.0
plotly>=6.5
